#393 feat: Move filters models to shared directory
