+++ src/pages/Glossary/models/index.ts
@@ -1,1 +1,1 @@
-export * from './filters.model';
+export * from './glossary-filters.model';
