+++ src/shared/models/index.ts
@@ -5,6 +5,7 @@ export * from './custom-parsed-query.model';
 export * from './sales-history.model';
 export * from './erc-1155-item.model';
 export * from './erc-1155-listings-batch.model';
+export * from './filters.model';
 export * from './installation.model';
 export * from './login-address.model';
 export * from './page-nav.model';
